requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
decorator==4.4.2
discord.py==1.7.1
dnspython==1.16.0
geographiclib==1.50
geopy==2.0.0
h11==0.11.0
httpcore==0.12.0
httpx==0.16.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
multidict==5.1.0
numpy==1.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
pyOpenSSL==20.0.1
python-binance==0.7.11
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.4.0
selenium==3.141.0
selenium-wire==2.1.1
service-identity==18.1.0
six==1.15.0
sniffio==1.2.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.25.11
watchdog==1.0.2
yarg==0.1.9
yarl==1.6.3
zope.interface==5.4.0
